pyproject.toml
[project]
name = "LlamaBot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai",
    "langchain_core",
    "python-dotenv",
    "langgraph",
    "langgraph_supervisor",
    "requests",
    "fastapi",
    "uvicorn",
    "starlette",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
pythonpath = ["."]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "websocket: marks tests as websocket related",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::RuntimeWarning",
]
addopts = "-v --tb=short" 